requirements.txt
Flask==2.3.3
boto3==1.28.85
python-dotenv==1.0.0
Werkzeug==2.3.7
textblob==0.17.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.3
botocore==1.31.85
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.18
nltk==3.8.1
regex==2023.10.3
joblib==1.3.2

# Advanced Features Dependencies
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
stripe==7.8.0
numpy==1.24.3
Pillow==10.0.1
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0

# AI and ML Dependencies
scikit-learn==1.3.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Security and Performance
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
cryptography==41.0.7
PyJWT==2.8.0

# Additional AWS Services
boto3[crt]==1.28.85
